pyproject.toml
[project]
name = "hdupy"
version = "0.1.0"
description = "A Python library for HDUers."
authors = [{ name = "hduLib" }]
dependencies = [
    "pycryptodome>=3.21.0",
    "lxml>=5.3.1",
    "yarl>=1.18.3",
    "typing-extensions>=4.12.2",
    "aiohttp>=3.11.13",
    "multidict>=6.1.0",
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = { text = "AGPL3" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.mypy]
follow_imports = "normal"
ignore_missing_imports = true
show_column_numbers = true
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_redundant_casts = true
disallow_untyped_calls = true
strict_equality = true

[tool.pdm]
distribution = true

[tool.pdm.build]
package-dir = "src"
includes = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.5",
    "pillow>=11.1.0",
    "types-lxml>=2025.3.4",
    "pyinstaller>=6.12.0",
    "matplotlib>=3.10.1",
]
